python/pytauri/pyproject.toml
[project]
name = "pytauri"
version = "0.2.0"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Sean Wang", email = "126865849+WSH032@users.noreply.github.com" },
]
requires-python = ">=3.9"
description = "Tauri binding for Python through Pyo3."

# <https://iscinumpy.dev/post/bound-version-constraints/>
dependencies = [
    "pydantic >= 2",
    "anyio >= 4",
    "pillow >= 11.1",         # `>=11` for free-threading
    "typing-extensions >= 4",
    # See: <https://pypi.org/project/backports.entry-points-selectable/>
    # and: <https://docs.python.org/3/library/importlib.metadata.html#entry-points>
    # Deprecated: once we no longer support versions Python 3.9, we can remove this dependency.
    "importlib-metadata >= 8",
    # workspaces, must use `==`
    # ...
]

[project.urls]
# see: <https://packaging.python.org/en/latest/specifications/well-known-project-urls/#well-known-labels>
documentation = "https://wsh032.github.io/pytauri/"
source = "https://github.com/WSH032/pytauri"


[tool.uv.sources]
# ...

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
